Feature: changed file names and added some styles to QR modal
