+++ src/components/GalleryItem/index.tsx
