--- src/components/QRModal/QRModal.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import './styles.scss'
-
-export const QRModal = () => {
-  return(
-    
-  )
-}
