+++ src/components/QRModal/index.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import QRCode from 'qrcode.react';
+import './styles.scss';
+
+export const QRModal = ({GeneratedUrl} : any) => {
+  return(
+    <section className="qrModal">
+      <h3>Here is your Pictcode</h3>
+      <QRCode value={GeneratedUrl}/>
+    </section>
+  )
+}
