+++ src/components/WorkspaceModal/styles.scss
