+++ src/index.tsx
@@ -1,6 +1,6 @@
 import ReactDOM from 'react-dom';
 import React from 'react';
 
-import App from './components/App/App';
+import App from './components/App/index';
 
 ReactDOM.render(<App/>, document.getElementById('root'));
